Fortify
Former-commit-id: 5c1912d911d51f8c288427f04e511dc88d5c66e0

+++ src/libcode/vx_data2d_grib/grib_classes.cc
@@ -1513,7 +1513,7 @@ double char4_to_dbl(const unsigned char *c)
 
    positive = (c[0] & 0x80) == 0;
    power = (int) (c[0] & 0x7f) - 64;
-   abspower = abs(power);
+   abspower = (unsigned int) (abs(power));
 
 
    /* calc exp */
